[PR] removed interface
