+++ src/main/java/io/tackle/pathfinder/controllers/AssessmentsResource.java
@@ -2,7 +2,10 @@ package io.tackle.pathfinder.controllers;
 
 import io.tackle.pathfinder.dto.ApplicationDto;
 import io.tackle.pathfinder.dto.AssessmentHeaderDto;
+import io.tackle.pathfinder.services.AssessmentSvc;
 
+import javax.inject.Inject;
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -10,24 +13,33 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
-/**
- * A JAX-RS interface.  An implementation of this interface must be provided.
- */
 @Path("/assessments")
-public interface AssessmentsResource {
-  /**
-   * Gets the details of a single instance of a `assessment`.
-   */
+public class AssessmentsResource {
+  @Inject
+  AssessmentSvc service;
+
   @GET
   @Produces("application/json")
-  public List<AssessmentHeaderDto> getApplicationAssessments(@NotNull @QueryParam("applicationId") Long applicationId);
+  public List<AssessmentHeaderDto> getApplicationAssessments(@NotNull @QueryParam("applicationId") Long applicationId) {
+    return service.getAssessmentHeaderDtoByApplicationId(applicationId).stream().collect(Collectors.toList());
+  }
 
   @POST
   @Produces("application/json")
   @Consumes("application/json")
-  Response createAssessment(@NotNull ApplicationDto data);
+  public Response createAssessment(@NotNull @Valid ApplicationDto data) {
+    return Response
+      .status(Status.CREATED)
+      .entity(service.createAssessment(data.getApplicationId()))
+      .type(MediaType.APPLICATION_JSON)
+      .build();
+  }
+
 }
